# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.RecommendRequest:
    properties:
      contextual:
        type: boolean
      direction:
        enum:
        - book_to_movie
        - book_to_book
        - movie_to_movie
        - movie_to_book
        type: string
      exclude_ids:
        items:
          type: string
        type: array
      selected_ids:
        items:
          type: string
        type: array
      weights:
        additionalProperties:
          format: float64
          type: number
        type: object
    required:
    - direction
    - selected_ids
    type: object
  handler.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  redis.RecommendationResult:
    properties:
      contextual:
        type: boolean
      created_at:
        type: integer
      error:
        type: string
      movies:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Cross-media recommendation system API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Recly API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: access_token, refresh_token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "409":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
  /book/{id}/like:
    delete:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Remove book from library
      tags:
      - library
    post:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Add book to library
      tags:
      - library
...
  /movie/{id}/like:
    delete:
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Remove movie from library
      tags:
      - library
    post:
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Add movie to library
      tags:
      - library
  /recommend:
    post:
      consumes:
      - application/json
      parameters:
      - description: Recommendation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RecommendRequest'
      produces:
      - application/json
      responses:
        "202":
          description: task_id, status
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Request recommendations
      tags:
      - recommendations
  /result/{taskId}:
    get:
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/redis.RecommendationResult'
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get recommendation result
      tags:
      - recommendations
...
  /user/library/books:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: books
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get liked books
      tags:
      - library
  /user/library/movies:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: movies
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get liked movies
      tags:
      - library
...
  /user/recommendations/history:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: history
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get recommendation history
      tags:
      - recommendations
  /user/saved-items:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: saved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get saved items
      tags:
      - saved
    post:
      consumes:
      - application/json
      parameters:
      - description: Item data (item_type, item_id)
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: saved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Save item
      tags:
      - saved
  /user/saved-items/{id}:
    delete:
      parameters:
      - description: Saved item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Delete saved item
      tags:
      - saved
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"